requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
pymongo==4.3.3
pytz==2022.7.1
sqlparse==0.2.4
